requirements.txt
Django==1.11.4
Markdown==2.6.8
certifi==2017.7.27.1
chardet==3.0.4	
django-annoying==0.10.3	
django-chartit==0.2.9	
django-filter==1.0.4	
django-model-utils==3.0.0
django-notifications-hq==1.3
django-webpack-loader==0.5.0
djangorestframework==3.6.3
idna==2.5
jsonfield==2.0.2
lxml==3.8.0	
mysqlclient==1.3.10	
npm==0.1.1	
optional-django==0.3.0	
pip==9.0.1	
pytz==2017.2	
requests==2.18.3	
setuptools==36.2.7	
sqlparse==0.2.3
urllib3==1.22
virtualenv==15.1.0	
webpack==6.0.0	